[test/rpi_yale] Broke testing script into two pieces.
